\name{SNP}
\alias{SNP}
\docType{data}
\title{
Simulated SNP signal
}
\description{
SNP profiles for 30 patients and 500 probes have been simulated. Each gene corresponds to one SNP value. 
}
\usage{data("SNP")}
\format{
  Data frame with 500 probes (rows) on 30 patients (patients).
}
\examples{
data(SNP)
}
\keyword{datasets}
